requirements.txt
altgraph==0.17.2
APScheduler==3.11.0
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
future==0.18.2
idna==3.7
macholib==1.15.2
pandas==2.2.3
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyupbit==0.2.34
requests==2.32.3
six==1.15.0
soupsieve==2.5
ta==0.11.0
tzdata==2024.2
tzlocal==5.2
urllib3==1.26.15
websockets==14.1
